consistent naming for GeoDataSource and api docs
