+++ hydromt/data_adapter/geodataframe.py
@@ -11,7 +11,6 @@ from pyproj import CRS
 from hydromt._typing import (
     Geom,
     NoDataStrategy,
-    TotalBounds,
 )
 from hydromt.gis import parse_geom_bbox_buffer, utils
 
@@ -181,64 +180,3 @@ class GeoDataFrameAdapter(DataAdapterBase):
                 gdf[col].attrs.update(**self.harmonization_settings.attrs[col])
 
         return gdf
-
-    def get_bbox(self, crs: Optional[CRS], detect: bool = True) -> TotalBounds:
-        """Return the bounding box and espg code of the dataset.
-
-        if the bounding box is not set and detect is True,
-        :py:meth:`hydromt.GeoDataframeAdapter.detect_bbox` will be used to detect it.
-
-        Parameters
-        ----------
-        detect: bool, Optional
-            whether to detect the bounding box if it is not set. If False, and it's not
-            set None will be returned.
-
-        Returns
-        -------
-        bbox: Tuple[np.float64,np.float64,np.float64,np.float64]
-            the bounding box coordinates of the data. coordinates are returned as
-            [xmin,ymin,xmax,ymax]
-        crs: int
-            The ESPG code of the CRS of the coordinates returned in bbox
-        """
-        bbox = self.harmonization_settings.extent.get("bbox", None)
-        if bbox is None and detect:
-            bbox, crs = self.detect_bbox()
-
-        return bbox, crs
-
-    # TODO: this should be done by the driver
-    def detect_bbox(
-        self,
-        gdf: Optional[gpd.GeoDataFrame] = None,
-    ) -> TotalBounds:
-        """Detect the bounding box and crs of the dataset.
-
-        If no dataset is provided, it will be fetched acodring to the settings in the
-        adapter. also see :py:meth:`hydromt.GeoDataframeAdapter.get_data`. the
-        coordinates are in the CRS of the dataset itself, which is also returned
-        alongside the coordinates.
-
-
-        Parameters
-        ----------
-        ds: xr.Dataset, xr.DataArray, Optional
-            the dataset to detect the bounding box of.
-            If none is provided, :py:meth:`hydromt.GeoDataframeAdapter.get_data`
-            will be used to fetch the it before detecting.
-
-        Returns
-        -------
-        bbox: Tuple[np.float64,np.float64,np.float64,np.float64]
-            the bounding box coordinates of the data. coordinates are returned as
-            [xmin,ymin,xmax,ymax]
-        crs: int
-            The ESPG code of the CRS of the coordinates returned in bbox
-        """
-        if gdf is None:
-            gdf = self.get_data()
-
-        crs = gdf.geometry.crs.to_epsg()
-        bounds = gdf.geometry.total_bounds
-        return bounds, crs
