+++ hydromt/data_adapter/rasterdataset.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import os
 from logging import Logger, getLogger
 from os.path import join
-from typing import Dict, List, Optional, Tuple, Union, cast
+from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
 import pandas as pd
@@ -25,7 +25,6 @@ from hydromt._typing import (
     RasterDatasetSource,
     StrPath,
     TimeRange,
-    TotalBounds,
     Variables,
     _exec_nodata_strat,
 )
@@ -559,122 +558,3 @@ class RasterDatasetAdapter(DataAdapterBase):
             return da
         else:
             return ds
-
-    def get_bbox(self, crs: Optional[CRS], detect=True) -> TotalBounds:
-        """Return the bounding box and espg code of the dataset.
-
-        if the bounding box is not set and detect is True,
-        :py:meth:`hydromt.RasterdatasetAdapter.detect_bbox` will be used to detect it.
-
-        Parameters
-        ----------
-        detect: bool, Optional
-            whether to detect the bounding box if it is not set. If False, and it's not
-            set None will be returned.
-
-        Returns
-        -------
-        bbox: Tuple[np.float64,np.float64,np.float64,np.float64]
-            the bounding box coordinates of the data. coordinates are returned as
-            [xmin,ymin,xmax,ymax]
-        crs: int
-            The ESPG code of the CRS of the coordinates returned in bbox
-        """
-        bbox = self.harmonization_settings.extent.get("bbox", None)
-        crs = cast(int, crs)
-        if bbox is None and detect:
-            bbox, crs = self.detect_bbox()
-
-        return bbox, crs
-
-    def get_time_range(
-        self,
-        detect=True,
-    ) -> TimeRange:
-        """Detect the time range of the dataset.
-
-        if the time range is not set and detect is True,
-        :py:meth:`hydromt.RasterdatasetAdapter.detect_time_range` will be used
-        to detect it.
-
-
-        Parameters
-        ----------
-        detect: bool, Optional
-            whether to detect the time range if it is not set. If False, and it's not
-            set None will be returned.
-
-        Returns
-        -------
-        range: Tuple[np.datetime64, np.datetime64]
-            A tuple containing the start and end of the time dimension. Range is
-            inclusive on both sides.
-        """
-        time_range = self.harmonization_settings.extent.get("time_range", None)
-        if time_range is None and detect:
-            time_range = self.detect_time_range()
-
-        return time_range
-
-    def detect_bbox(
-        self,
-        ds=None,
-    ) -> TotalBounds:
-        """Detect the bounding box and crs of the dataset.
-
-        If no dataset is provided, it will be fetched according to the settings in the
-        adapter. also see :py:meth:`hydromt.RasterdatasetAdapter.get_data`. the
-        coordinates are in the CRS of the dataset itself, which is also returned
-        alongside the coordinates.
-
-
-        Parameters
-        ----------
-        ds: xr.Dataset, xr.DataArray, Optional
-            the dataset to detect the bounding box of.
-            If none is provided, :py:meth:`hydromt.RasterdatasetAdapter.get_data`
-            will be used to fetch the it before detecting.
-
-        Returns
-        -------
-        bbox: Tuple[np.float64,np.float64,np.float64,np.float64]
-            the bounding box coordinates of the data. coordinates are returned as
-            [xmin,ymin,xmax,ymax]
-        crs: int
-            The ESPG code of the CRS of the coordinates returned in bbox
-        """
-        if ds is None:  # TODO: If we keep this optional, should move to DataSource.
-            ds = self.get_data()
-        crs = ds.raster.crs.to_epsg()
-        bounds = ds.raster.bounds
-
-        return bounds, crs
-
-    def detect_time_range(
-        self,
-        ds=None,
-    ) -> TimeRange:
-        """Detect the temporal range of the dataset.
-
-        If no dataset is provided, it will be fetched accodring to the settings in the
-        addapter. also see :py:meth:`hydromt.RasterdatasetAdapter.get_data`.
-
-        Parameters
-        ----------
-        ds: xr.Dataset, xr.DataArray, Optional
-            the dataset to detect the time range of. It must have a time dimentsion set.
-            If none is provided, :py:meth:`hydromt.RasterdatasetAdapter.get_data`
-            will be used to fetch the it before detecting.
-
-        Returns
-        -------
-        range: Tuple[np.datetime64, np.datetime64]
-            A tuple containing the start and end of the time dimension. Range is
-            inclusive on both sides.
-        """
-        if ds is None:  # TODO: If we want to keep this optional, should move to Source.
-            ds = self.get_data()
-        return (
-            ds[ds.raster.time_dim].min().values,
-            ds[ds.raster.time_dim].max().values,
-        )
